# Made sure inference completes in _inference, no lazy evaluation