*Issue #, if available:*
#414 
*Description of changes:*
Complete inference in _inference

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.