@vrakesh 2018-05-24
With change (on squeezenet)

LatencyInference for last 60 seconds is 42.354107
[INFO 2018-05-24 14:45:37,767 PID:33812 /Users/rakvas/Workspace/mxnet-model-server/mms/metric.py:start_recording:216] Metric LatencyPostprocess for last 60 seconds is 5.860090

Before change
LatencyInference for last 60 seconds is 0.210047
Metric LatencyPostprocess for last 60 seconds is 47.446966