@vrakesh 2018-05-24
With the post process step, removed the latency after change for post process is 0.09ms. The post process step with top_k for squeezenet does have about 5ms latency still